Separated actions into its own router for readability
